backend/lambda/requirements.txt
google-generativeai==0.8.3
pydantic==2.10.6
boto3==1.34.0